+++ src/AiServices/OpenAi.php
@@ -7,7 +7,7 @@ namespace MageOS\AiBase\AiServices;
 use MageOS\AiBase\Api\Data\AiServiceConfigurationInterface;
 use MageOS\AiBase\Api\Data\FieldDescriptorInterfaceFactory;
 
-final class OpenAi implements AiServiceConfigurationInterface
+class OpenAi implements AiServiceConfigurationInterface
 {
     use FieldFactoryTrait;